added dimming for connect/disconnect
mostly implements #28
there is a need for a trailing preamble still to do but it needs a coordinated effort between the windows code and the arduino code. if the trailing preamble (=postamble?) is not used, it is impossible to decide whether the serial connection was lost or the screen is just black!